requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
loguru==0.5.3
pynetbox==6.1.2
python-dotenv==0.18.0
requests==2.25.1
six==1.16.0
typing==3.7.4.3
urllib3==1.26.5
